% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{base64ToHtml}
\alias{base64ToHtml}
\title{Base64 str to Html page}
\usage{
base64ToHtml(base64Chart = makeFieldChart())
}
\arguments{
\item{base64Chart}{Base64 문자열}
}
\value{
Html page.
}
\description{
base64 문자열을 html 페이지로 띄워서 보기 위한 함수
}
