% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{checkSignal}
\alias{checkSignal}
\title{Check Signal}
\usage{
checkSignal(df, target, type)
}
\arguments{
\item{df}{group, x축 좌표, y축 좌표 구조로 대상 그룹으로 Filtering 후 입력 필요}

\item{target}{목표값으로 L3M의 경우 전년 실적 입력 필요}

\item{type}{시그널을 확인하고자 하는 차트 종류 ffr, L6M 또는 L3M}
}
\value{
color str
}
\description{
시그널 확인 함수
}
