% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{checkWeekSignal}
\alias{checkWeekSignal}
\title{Check Week Signal}
\usage{
checkWeekSignal(last = 0, this = 0)
}
\arguments{
\item{last}{지난주 실적, 기본값 = 0}

\item{this}{금주 실적, 기본값 = 0}
}
\value{
color str
}
\description{
지난주 금주 실적 시그널 확인 함수
}
