% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{makeFieldChart}
\alias{makeFieldChart}
\title{Make a field chart from df}
\usage{
makeFieldChart(
  wd = getwd(),
  df = dxChart::ffr_fdr_sample,
  yCol = "value",
  xCol = "PURC_MON_NEW",
  groupCol = "group",
  xType = "datetime",
  xLeftMargin = 0.15,
  yMaxRate = 1.4,
  yLeftText = "FFR(\%)",
  yRightText = "FDR(\%)",
  lineWidth = 1,
  tickIntervalY = 0.5,
  tickIntervalX = 30 * 24 * 3600 * 1000,
  useCustomize = TRUE,
  linelabelSignals = c("red", "green", "grey", "green", "red"),
  linelabelSymbols = c("", "", "●", "", "●"),
  lineSymbols = c("circle", "circle", "diamond", "diamond", "square"),
  lineSymbolColors = c("white", "", "", "", "white"),
  markerHover = TRUE,
  groupColors = c("#000000", "#FF0000", "#FF00FF", "#7F7F7F", "#FFC000"),
  useDatalabels = c(TRUE, TRUE, TRUE, FALSE, TRUE),
  yRightUse = TRUE,
  useLeftlabels = TRUE,
  useLinelabels = FALSE,
  useWeeklabels = TRUE,
  fontFamily = "LG스마트체 Regular",
  titleText = "Global OLED (Product)",
  titleFontWeight = "bold",
  titleFontSize = "16px",
  linelabelFontWeight = "bold",
  linelabelFontSize = "12px",
  weeklabelFontWeight = "bold",
  weeklabelFontSize = "12px",
  datalabelFontWeight = "normal",
  datalabelOutline = "1px white",
  imageHeight = 400,
  imageWidth = 640,
  base64 = TRUE
)
}
\arguments{
\item{wd}{작업 디렉토리, 기본값 = getwd()}

\item{df}{데이터프레임(group, x축 좌표(문자), y축 좌표(숫자), 기본값 = dxChart::ffr_fdr_sample}

\item{yCol}{y축 좌표가 위치한 행, 기본값 = "value"}

\item{xCol}{x축 좌표가 위치한 행, 기본값 = "PURC_MON_NEW"}

\item{groupCol}{데이터 Group이 위치한 행, 기본값 = "group"}

\item{xType}{x축의 타입 : "datetime" 또는 "category", 기본값 = "datetime"}

\item{xLeftMargin}{x축 좌측의 여백값으로 타입에 따라 값을 변경 해야함, 기본값 = 0.15}

\item{yMaxRate}{y축 좌표의 최대값 대비 y축의 최대값 비율, 기본값 = 1.4}

\item{yLeftText}{y축 좌측 문구, 기본값 = "FFR(\%)"}

\item{yRightText}{y축 우측 문구, 기본값 = "FDR(\%)"}

\item{lineWidth}{라인 두께, 기본값 = 1}

\item{tickIntervalY}{y축 라벨 표기 간격, 기본값 = 0.5}

\item{tickIntervalX}{x축 라벨 표기 간격으로 datetime 타입의 경우 초단위로 설정,
기본값 = 30 * 24 * 3600 * 1000}

\item{linelabelSignals}{라인라벨의 시그널 색상,
기본값 = c("red", "green", "grey", "green", "red"),}

\item{linelabelSymbols}{라인라벨 시그널의 모양, 기본값 = c("", "", "●", "", "●"),}

\item{lineSymbols}{라인의 심볼 : circle, diamond 또는 square ,
기본값 = c('circle', 'circle', 'diamond', 'diamond', 'square'),}

\item{lineSymbolColors}{라인 심볼의 색상으로 설정하지 않으면 라인 색상을 따라감,
기본값 = c('white', '', '', '', 'white'),}

\item{markerHover}{라인에 마우스를 올렸을 때 라인 심볼의 표시 여부 : TRUE 또는 FALSE, 기본값 = TRUE}

\item{groupColors}{그룹별 라인과 라인라벨의 색상 : 색상코드 또는 FALSE,
기본값 = c("#000000", "#FF0000", "#FF00FF", "#7F7F7F", "#FFC000")}

\item{useDatalabels}{데이터 라벨의 표시 여부 : TRUE 또는 FALSE, 기본값 = c(TRUE, TRUE, TRUE, FALSE, TRUE),}

\item{yRightUse}{우측 Y축을 사용할지 여부, 기본값 = TRUE}

\item{useLeftlabels}{좌측 라벨을 사용할지 여부, 기본값 = TRUE}

\item{useLinelabels}{라인별 라벨을 사용할지 여부, 기본값 = FALSE}

\item{useWeeklabels}{주간 라벨을 사용할지 여부, 기본값 = TRUE}

\item{fontFamily}{타이틀, 라벨, 데이터라벨의 폰트, 기본값 = "LG스마트체 Regular"}

\item{titleText}{타이틀 문구, 기본값 = "Global OLED (Product)"}

\item{titleFontWeight}{타이틀 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{titleFontSize}{타이틀 폰트 사이즈, 기본값 = "16px"}

\item{linelabelFontWeight}{라인라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{linelabelFontSize}{라인라벨의 폰트 사이즈, 기본값 = "12px"}

\item{weeklabelFontWeight}{주간 라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{weeklabelFontSize}{주간 라벨의 폰트 사이즈, 기본값 = "12px"}

\item{datalabelFontWeight}{데이터 라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = "normal"}

\item{datalabelOutline}{데이터 라벨의 아웃라인 : "사이즈 색상", 기본값 = "1px white"}

\item{imageHeight}{base64 이미지의 높이, 기본값 = 400}

\item{imageWidth}{base64 이미지의 넓이, 기본값 = 640}

\item{base64}{base64 이미지 또는 htmlwidget object 출력을 선택, 기본값 = TRUE}
}
\value{
base64 문자열 또는 htmlwidget object.
}
\description{
Field Chart를 만들기 위한 함수
}
