% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{makePresetChart}
\alias{makePresetChart}
\title{Make Chart by preset}
\usage{
makePresetChart(preset = "ffr", df = NaN, title = "", base64 = TRUE)
}
\arguments{
\item{preset}{ffr, hazard}

\item{df}{데이터프레임(group, x축 좌표(문자), y축 좌표(숫자), 기본값 = NULL}

\item{title}{타이틀 문구, 기본값 = ""}

\item{base64}{base64 이미지 또는 htmlwidget object 출력을 선택, 기본값 = TRUE}
}
\value{
base64 str or htmlwidget object
}
\description{
Preset Argument를 넣어서 차트를 만드는 함수
}
