% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{makeSampleChart}
\alias{makeSampleChart}
\title{Make Sample Chart}
\usage{
makeSampleChart(type = "ffr")
}
\arguments{
\item{type}{ffr 또는 hazard}
}
\value{
htmlwidget object
}
\description{
Type별 Sample 데이터를 넣어서 차트를 만드는 함수
}
