% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryFieldData}
\alias{queryFieldData}
\title{query field data from DB}
\usage{
queryFieldData(
  date,
  period = 2,
  type = "prod",
  cal = "ffr",
  accNum = 6,
  svcFilters,
  sellinFilters
)
}
\arguments{
\item{date}{검색 기준일, Datetime(yyyy-mm-dd)}

\item{period}{검색 기간(년), Int}

\item{type}{생산/판매 법인 기준 선택, prod or sales}

\item{cal}{계산 기준, ffr or fdr}

\item{accNum}{누적 기간으로 fdr를 계산 시 사용, Int}

\item{svcFilters}{SVC 데이터 쿼리의 Where문에 삽인되는 필터, list()}

\item{sellinFilters}{Sellin 데이터 쿼리의 Where문에 삽인되는 필터, list()}
}
\value{
svc 수량, sales 수량, 불량율
}
\description{
Field 불량율을 계산하는 함수
}
