% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{makeFieldChart}
\alias{makeFieldChart}
\title{Make a field chart from df}
\usage{
makeFieldChart(
  wd = getwd(),
  df = dxChart::ffr_fdr_sample,
  yCol = "value",
  xCol = "PURC_MON_NEW",
  xType = "datatime",
  xLeftMargin = 0.15,
  yMaxRate = 1.4,
  yLeftText = "FFR(\%)",
  yRightText = "FDR(\%)",
  lineWidth = 1,
  tickIntervalY = 0.5,
  tickIntervalX = 30 * 24 * 3600 * 1000,
  fontFamily = "LG스마트체 Regular",
  titleText = "Global OLED (Product)",
  titleFontWeight = "bold",
  titleFontSize = "16px",
  linelabelFontWeight = "bold",
  linelabelFontSize = "12px",
  weeklabelFontWeight = "bold",
  weeklabelFontSize = "12px",
  useDatalabels = TRUE,
  datalabelFontWeight = "normal",
  datalabelOutline = "1px white",
  imageHeight = 400,
  imageWidth = 640,
  base64 = TRUE
)
}
\arguments{
\item{wd}{작업 디렉토리, 기본값 = getwd()}

\item{df}{데이터프레임(group, x축 좌표(문자), y축 좌표(숫자), 기본값 = read.csv("./data/sample.csv", header = T)}

\item{yCol}{y축 좌표가 위치한 행, 기본값 = "value"}

\item{xCol}{x축 좌표가 위한 행, 기본값 = "PURC_MON_NEW"}

\item{xType}{x축의 타입 : "datetime" 또는 "category", 기본값 = "datatime"}

\item{xLeftMargin}{x축 좌측의 여백값으로 타입에 따라 값을 변경 해야함, 기본값 = 0.15}

\item{yMaxRate}{y축 좌표의 최대값 대비 y축의 최대값 비율, 기본값 = 1.4}

\item{yLeftText}{y축 좌측 문구, 기본값 = "FFR(\%)"}

\item{yRightText}{y축 우측 문구, 기본값 = "FDR(\%)"}

\item{lineWidth}{라인 두께, 기본값 = 1}

\item{tickIntervalY}{y축 라벨 표기 간격, 기본값 = 0.5}

\item{tickIntervalX}{x축 라벨 표기 간격으로 datetime 타입의 경우 초단위로 설정, 기본값 = 30 * 24 * 3600 * 1000}

\item{fontFamily}{타이틀, 라벨, 데이터라벨의 폰트, 기본값 = "LG스마트체 Regular"}

\item{titleText}{타이틀 문구, 기본값 = "Global OLED (Product)"}

\item{titleFontWeight}{타이틀 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{titleFontSize}{타이틀 폰트 사이즈, 기본값 = "16px"}

\item{linelabelFontWeight}{라인라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{linelabelFontSize}{라인라벨의 폰트 사이즈, 기본값 = "12px"}

\item{weeklabelFontWeight}{주간 라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = 'bold'}

\item{weeklabelFontSize}{주간 라벨의 폰트 사이즈, 기본값 = "12px"}

\item{useDatalabels}{데이터 라벨의 사용 여부, 기본값 = TRUE}

\item{datalabelFontWeight}{데이터 라벨의 폰트 스타일 : 'normal' 또는 bold', 기본값 = "normal"}

\item{datalabelOutline}{데이터 라벨의 아웃라인 : "사이즈 색상", 기본값 = "1px white"}

\item{imageHeight}{base64 이미지의 높이, 기본값 = 400}

\item{imageWidth}{base64 이미지의 넓이, 기본값 = 640}

\item{base64}{base64 이미지 또는 htmlwidget object 출력을 선택, 기본값 = TRUE}
}
\value{
base64 문자열 또는 htmlwidget object.
}
\description{
Field Chart를 만들기 위한 함수
}
